[NOTIFY] Send notification emails on reactions
When a user reacts to a movie send a notification email. For testing
emails can be viewed via normal rails mailer preview and wil be sent to
local mailcatcher for checking. For production use Sendgrid is
preconfigured and needs appropriately configured USERNAME and PASSWORD
provided via env variables, see production.rb for configuration.